feat: Add kingdom creation free boosts
Introduces a new `creationFreeBoosts` property to the Kingdom type and context. This allows players to select two free attribute boosts during kingdom creation, similar to character creation in Pathfinder 2e. The UI in `KingdomSheet` has been updated to reflect this new selection. The initial value for `creationFreeBoosts` is set to `[null, null]` in `constants.ts`.